Modify: filename and variable name of clock.py changed to make the meaning clear
